+++ CHANGELOG.md
@@ -12,8 +12,13 @@ Since 2018.1, the version numbers and release cycle match Rider's versions and r
 * [Commits](https://github.com/JetBrains/resharper-unity/compare/net212-rtm-2021.2.0...net212)
 * [Milestone](https://github.com/JetBrains/resharper-unity/milestone/49?closed=1)
 
+### Changed
+
+- Methods marked with `[UnityTearDown]` are treated as in use ([#2135](https://github.com/JetBrains/resharper-unity/pull/2135))
+
 ### Fixed
 
+- Rider: Builtin packages are correctly resolved against the local package cache for newer Unity versions (RIDER-65196](https://youtrack.jetbrains.com/issue/RIDER-65196), [#2122](https://github.com/JetBrains/resharper-unity/issues/2122), [#2136](https://github.com/JetBrains/resharper-unity/pull/2136))
 - Rider: Support "Show in Unity" action for files from a local package ([RIDER-65128](https://youtrack.jetbrains.com/issue/RIDER-65128), [#2124](https://github.com/JetBrains/resharper-unity/pull/2124))
 - Rider: Improve memory management when opening massive shader files ([RIDER-65080](https://youtrack.jetbrains.com/issue/RIDER-65080), [#2121](https://github.com/JetBrains/resharper-unity/pull/2121))
 - Rider: Fix UnityYAMLMerge failing to resolve conflicts in assets ([RIDER-64433](https://youtrack.jetbrains.com/issue/RIDER-64433), [#2125](https://github.com/JetBrains/resharper-unity/pull/2125))